third_party/blink/web_tests/fast/canvas/canvas-lost-gpu-context.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/canvas-context-lost-events.js"></script>
</head>
<body>
<script>
var ctx;
var lostEventHasFired = false;
var contextLostTest;

function contextLostWrapper() {
    contextLost(lostEventHasFired, ctx);
}

function contextRestoredWrapper() {
    contextRestored(lostEventHasFired, ctx);
}

async_test(function(t) {
    if (window.internals) {
        var canvas = document.createElement('canvas');
        canvas.addEventListener('contextlost', contextLostWrapper);
        canvas.addEventListener('contextrestored', contextRestoredWrapper);
        ctx = canvas.getContext('2d');
        document.body.appendChild(ctx.canvas);

        ctx.fillRect(0, 0, 1, 1);
        // setTimeout creates a frame barrier that locks the canvas into gpu
        // acceleration mode when running under virtual/gpu
        setTimeout(t.step_func_done(() => {
            // Now it is safe to use verifyContextLost without fearing side-effects
            // because a rendering mode was fixed.
            verifyContextLost(false, ctx);

            internals.forceLoseCanvasContext(canvas, "2d");
            if (!ctx.isContextLost()) {
                assert_true(false, "canvas context is not lost properly.");
            } else {
                verifyContextLost(true, ctx);
            }
        }), 0);
    } else {
        assert_true(false, "This test requires window.internals.");
    }
}, "Test the behavior of disconneced canvas recovery after a gpu context loss");

</script>
</body>
</html>
